Move v5 Slider to "Legacy" export
