+++ components/slider/legacy-styled.jsx
