+++ index.js
@@ -32,8 +32,8 @@ export {
 	Tooltip,
 } from './components/popover';
 export { Radio } from './components/radio';
-export { Slider } from './components/slider';
 export { SimpleToast } from './components/simple-toast';
+export { LegacySlider as Slider } from './components/slider';
 export {
 	TabManager,
 	Tab,
